Move overrides API to httpclient
